requirements.txt
textual
python-dotenv
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
pydantic
httpx
requests
zod
click
rich
py-clob-client
pytest